+++ src/imgui/ImGuiMedia.cc
@@ -1436,7 +1436,7 @@ void ImGuiMedia::cassetteMenu(CassettePlayer& cassettePlayer)
 			const auto now = motherBoard->getCurrentTime();
 			auto length = cassettePlayer.getTapeLength(now);
 			auto pos = cassettePlayer.getTapePos(now);
-			auto format = [](float time) {
+			auto format = [](double time) {
 				int t = narrow_cast<int>(time); // truncated to seconds
 				int s = t % 60; t /= 60;
 				int m = t % 60; t /= 60;
